+++ lib/idx/types/index.ts
@@ -116,6 +116,7 @@ export interface IdxTransaction {
   meta?: IdxTransactionMeta;
   enabledFeatures?: IdxFeature[];
   availableSteps?: NextStep[];
+  requestDidSucceed?: boolean;
 
   // from idx-js, used by signin widget
   proceed: (remediationName: string, params: unknown) => Promise<IdxResponse>;
@@ -147,4 +148,12 @@ export type Authenticator = {
 
 export function isAuthenticator(obj: any): obj is Authenticator {
   return obj && (obj.key || obj.id);
+}
+
+export interface RemediationResponse {
+  idxResponse: IdxResponse;
+  nextStep?: NextStep;
+  messages?: IdxMessage[];
+  terminal?: boolean;
+  canceled?: boolean;
 }